+++ src/components/Advocacy.vue
@@ -90,7 +90,7 @@ li{
   padding: 0.5rem;
 }
 button{
-  background: #E7612C;
+  background: #f95607;
   color: white;
   border-radius: 5px;
   border: 2px solid  #E7612C;
